+++ src/main/java/br/gov/es/openpmo/controller/permissions/OfficePermissionController.java
@@ -1,12 +1,13 @@
 package br.gov.es.openpmo.controller.permissions;
 
+import br.gov.es.openpmo.configuration.Authorization;
 import br.gov.es.openpmo.dto.ResponseBase;
 import br.gov.es.openpmo.dto.officepermission.OfficePermissionDto;
 import br.gov.es.openpmo.dto.officepermission.OfficePermissionParamDto;
 import br.gov.es.openpmo.model.Entity;
 import br.gov.es.openpmo.service.authentication.TokenService;
 import br.gov.es.openpmo.service.permissions.OfficePermissionService;
-import br.gov.es.openpmo.utils.ApplicationMessage;
+import br.gov.es.openpmo.service.permissions.canaccess.ICanAccessService;
 import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -23,14 +24,17 @@ public class OfficePermissionController {
 
   private final OfficePermissionService service;
   private final TokenService tokenService;
+  private final ICanAccessService canAccessService;
 
   @Autowired
   public OfficePermissionController(
-    OfficePermissionService service,
-    TokenService tokenService
+    final OfficePermissionService service,
+    final TokenService tokenService,
+    final ICanAccessService canAccessService
   ) {
     this.service = service;
     this.tokenService = tokenService;
+    this.canAccessService = canAccessService;
   }
 
   @GetMapping
@@ -38,51 +42,58 @@ public class OfficePermissionController {
     @RequestParam(name = "id-office") final Long idOffice,
     @RequestParam(required = false) final Long idFilter,
     @RequestParam(name = "key", required = false) final String key,
-    @RequestHeader(name = "Authorization") final String authorization
+    @Authorization final String authorization
   ) {
+    this.canAccessService.ensureCanReadManagementResource(idOffice, key, authorization);
     final Long idPerson = this.tokenService.getUserId(authorization);
     final List<OfficePermissionDto> offices = this.service.findAllDto(idOffice, idFilter, key, idPerson);
-    if (offices.isEmpty()) {
+    if(offices.isEmpty()) {
       return ResponseEntity.noContent().build();
     }
-    final ResponseBase<List<OfficePermissionDto>> response = new ResponseBase<List<OfficePermissionDto>>()
-      .setData(offices).setSuccess(true).setMessage(ApplicationMessage.OPERATION_SUCCESS);
-    return ResponseEntity.ok(response);
+    return ResponseEntity.ok(ResponseBase.of(offices));
   }
 
   @GetMapping("/{id-office}")
   public ResponseEntity<ResponseBase<OfficePermissionDto>> findPermissionsById(
     @PathVariable(name = "id-office") final Long idOffice,
     @RequestParam(name = "key", required = false) final String key,
-    @RequestHeader(name = "Authorization") final String authorization
+    @Authorization final String authorization
   ) {
-    Long idPerson = this.tokenService.getUserId(authorization);
-    OfficePermissionDto permissions = this.service.findOfficePermissionsByKey(idOffice, key, idPerson);
+    this.canAccessService.ensureCanReadManagementResource(idOffice, key, authorization);
+    final Long idPerson = this.tokenService.getUserId(authorization);
+    final OfficePermissionDto permissions = this.service.findOfficePermissionsByKey(idOffice, key, idPerson);
     return ResponseEntity.ok(ResponseBase.of(permissions));
   }
 
   @PostMapping
-  public ResponseEntity<ResponseBase<Entity>> store(@Valid @RequestBody final OfficePermissionParamDto request) {
+  public ResponseEntity<ResponseBase<Entity>> store(
+    @Valid @RequestBody final OfficePermissionParamDto request,
+    @Authorization final String authorization
+  ) {
+    this.canAccessService.ensureCanAccessManagementResource(request.getIdOffice(), authorization);
     this.service.store(request);
-    final ResponseBase<Entity> response = new ResponseBase<Entity>()
-      .setSuccess(true)
-      .setMessage(ApplicationMessage.OPERATION_SUCCESS);
-    return ResponseEntity.ok(response);
+    return ResponseEntity.ok(ResponseBase.of());
   }
 
   @PutMapping
-  public ResponseEntity<ResponseBase<Entity>> update(@Valid @RequestBody final OfficePermissionParamDto request) {
+  public ResponseEntity<ResponseBase<Entity>> update(
+    @Valid @RequestBody final OfficePermissionParamDto request,
+    @Authorization final String authorization
+  ) {
+    this.canAccessService.ensureCanAccessManagementResource(request.getIdOffice(), authorization);
     this.service.update(request);
-    final ResponseBase<Entity> entity = new ResponseBase<Entity>().setSuccess(true).setMessage(ApplicationMessage.OPERATION_SUCCESS);
-    return ResponseEntity.ok(entity);
+    return ResponseEntity.ok(ResponseBase.of());
   }
 
   @DeleteMapping
   public ResponseEntity<Void> delete(
     @RequestParam(name = "id-office") final Long idOffice,
-    @RequestParam(name = "key") final String key
+    @RequestParam(name = "key") final String key,
+    @Authorization final String authorization
   ) {
+    this.canAccessService.ensureCanAccessManagementResource(idOffice, authorization);
     this.service.delete(idOffice, key);
     return ResponseEntity.ok().build();
   }
+
 }
